+++ tests/string/Base64.test.ts
