+++ tests/string/Hexadecimal.test.ts
@@ -3,20 +3,39 @@ import * as E from 'fp-ts/Either'
 
 import { pipe, tuple } from 'fp-ts/function'
 
-import * as HexColor from '../src/string/HexColor'
+import * as Hexadecimal from '../../src/string/Hexadecimal'
 
-import { cat, combineExpected, validateArbitrary } from '../test-utils'
+import { cat, combineExpected, validateArbitrary } from '../../test-utils'
 
-const validStrings = ['#ff0000ff', '#ff0034', '#CCCCCC', '0f38', 'fff', '#f00']
+const validStrings = [
+  'deadBEEF',
+  'ff0044',
+  '0xff0044',
+  '0XfF0044',
+  '0x0123456789abcDEF',
+  '0X0123456789abcDEF',
+  '0hfedCBA9876543210',
+  '0HfedCBA9876543210',
+  '0123456789abcDEF',
+]
 
-const invalidStrings = ['#ff', 'fff0a', '#ff12FG', '#bbh']
+const invalidStrings = [
+  'abcdefg',
+  '',
+  '..',
+  '0xa2h',
+  '0xa20x',
+  '0x0123456789abcDEFq',
+  '0hfedCBA9876543210q',
+  '01234q56789abcDEF',
+]
 
-describe('HexColor', () => {
+describe('Hexadecimal', () => {
   describe('Decoder', () => {
     test.each(
       cat(combineExpected(validStrings, 'Right'), combineExpected(invalidStrings, 'Left'))
     )('validates valid strings, and catches bad strings', (str, expectedTag) => {
-      const result = HexColor.Decoder.decode(str)
+      const result = Hexadecimal.Decoder.decode(str)
 
       expect(result._tag).toBe(expectedTag)
     })
@@ -28,8 +47,8 @@ describe('HexColor', () => {
       original => {
         const roundtrip = pipe(
           original,
-          HexColor.Decoder.decode,
-          E.map(HexColor.Encoder.encode),
+          Hexadecimal.Decoder.decode,
+          E.map(Hexadecimal.Encoder.encode),
           E.getOrElse(() => 'invalid')
         )
 
@@ -43,8 +62,8 @@ describe('HexColor', () => {
       'determines two strings are equal',
 
       (str1, str2) => {
-        const guard = HexColor.Guard.is
-        const eq = HexColor.Eq.equals
+        const guard = Hexadecimal.Guard.is
+        const eq = Hexadecimal.Eq.equals
 
         if (!guard(str1) || !guard(str2)) {
           throw new Error('Unexpected result')
@@ -59,7 +78,7 @@ describe('HexColor', () => {
     test.each(
       cat(combineExpected(validStrings, true), combineExpected(invalidStrings, false))
     )('validates valid strings, and catches bad strings', (str, expectedTag) => {
-      const result = HexColor.Guard.is(str)
+      const result = Hexadecimal.Guard.is(str)
 
       expect(result).toBe(expectedTag)
     })
@@ -69,7 +88,7 @@ describe('HexColor', () => {
     test.each(
       cat(combineExpected(validStrings, 'Right'), combineExpected(invalidStrings, 'Left'))
     )('validates valid strings, and catches bad strings', async (str, expectedTag) => {
-      const result = await HexColor.TaskDecoder.decode(str)()
+      const result = await Hexadecimal.TaskDecoder.decode(str)()
 
       expect(result._tag).toBe(expectedTag)
     })
@@ -79,15 +98,15 @@ describe('HexColor', () => {
     test.each(
       cat(combineExpected(validStrings, 'Right'), combineExpected(invalidStrings, 'Left'))
     )('validates valid strings, and catches bad strings', (str, expectedTag) => {
-      const result = HexColor.Type.decode(str)
+      const result = Hexadecimal.Type.decode(str)
 
       expect(result._tag).toBe(expectedTag)
     })
   })
 
   describe('Arbitrary', () => {
-    it('generates valid HexColors', () => {
-      validateArbitrary(HexColor, HexColor.isHexColor)
+    it('generates valid Hexadecimals', () => {
+      validateArbitrary(Hexadecimal, Hexadecimal.isHexadecimal)
     })
   })
 })
