+++ tests/string/PositiveFloatString.test.ts
@@ -2,32 +2,32 @@ import * as RA from 'fp-ts/ReadonlyArray'
 
 import { tuple } from 'fp-ts/function'
 
-import * as NegativeFloatString from '../src/string/NegativeFloatString'
+import * as PositiveFloatString from '../../src/string/PositiveFloatString'
 
-import { cat, combineExpected, validateArbitrary } from '../test-utils'
+import { cat, combineExpected, validateArbitrary } from '../../test-utils'
 
-const validNumbers = ['-1', '-1.1', `-${Math.random() + 1}`, `${Number.MIN_SAFE_INTEGER}`]
+const validNumbers = ['1', '1.1', `${Math.random() + 1}`, `${Number.MAX_SAFE_INTEGER}`]
 
 const invalidNumbers = [
   '0',
-  '1',
+  '-1',
   '1.1.1.1',
   '2......',
   'a',
-  '1.1',
-  `${Math.random()}`,
-  `${Number.MAX_SAFE_INTEGER}`,
+  '-1.1',
+  `${-Math.random()}`,
+  `${Number.MIN_SAFE_INTEGER}`,
   `${Infinity}`,
   `${-Infinity}`,
   `${NaN}`,
 ]
 
-describe('NegativeFloatString', () => {
+describe('PositiveFloatString', () => {
   describe('Decoder', () => {
     test.each(
       cat(combineExpected(validNumbers, 'Right'), combineExpected(invalidNumbers, 'Left'))
     )('validates valid numbers, and catches bad numbers', (num, expectedTag) => {
-      const result = NegativeFloatString.Decoder.decode(num)
+      const result = PositiveFloatString.Decoder.decode(num)
 
       expect(result._tag).toBe(expectedTag)
     })
@@ -38,10 +38,10 @@ describe('NegativeFloatString', () => {
       'determines two numbers are equal',
 
       (num1, num2) => {
-        if (!NegativeFloatString.Guard.is(num1) || !NegativeFloatString.Guard.is(num2))
+        if (!PositiveFloatString.Guard.is(num1) || !PositiveFloatString.Guard.is(num2))
           throw new Error('Unexpected result')
 
-        expect(NegativeFloatString.Eq.equals(num1, num2)).toBe(true)
+        expect(PositiveFloatString.Eq.equals(num1, num2)).toBe(true)
       }
     )
   })
@@ -50,7 +50,7 @@ describe('NegativeFloatString', () => {
     test.each(
       cat(combineExpected(validNumbers, true), combineExpected(invalidNumbers, false))
     )('validates valid numbers, and catches bad numbers', (num, expectedTag) => {
-      const result = NegativeFloatString.Guard.is(num)
+      const result = PositiveFloatString.Guard.is(num)
 
       expect(result).toBe(expectedTag)
     })
@@ -60,7 +60,7 @@ describe('NegativeFloatString', () => {
     test.each(
       cat(combineExpected(validNumbers, 'Right'), combineExpected(invalidNumbers, 'Left'))
     )('validates valid numbers, and catches bad numbers', async (num, expectedTag) => {
-      const result = await NegativeFloatString.TaskDecoder.decode(num)()
+      const result = await PositiveFloatString.TaskDecoder.decode(num)()
 
       expect(result._tag).toBe(expectedTag)
     })
@@ -70,20 +70,20 @@ describe('NegativeFloatString', () => {
     test.each(
       cat(combineExpected(validNumbers, 'Right'), combineExpected(invalidNumbers, 'Left'))
     )('validates valid numbers, and catches bad numbers', (num, expectedTag) => {
-      const result = NegativeFloatString.Type.decode(num)
+      const result = PositiveFloatString.Type.decode(num)
 
       expect(result._tag).toBe(expectedTag)
     })
   })
   it('converts to a PositiveFloat', () => {
-    const numStr = '-1.1'
-    if (!NegativeFloatString.Guard.is(numStr)) throw new Error('Unexpected result')
-    expect(NegativeFloatString.toNegativeFloat(numStr)).toBe(-1.1)
+    const numStr = '1.1'
+    if (!PositiveFloatString.Guard.is(numStr)) throw new Error('Unexpected result')
+    expect(PositiveFloatString.toPositiveFloat(numStr)).toBe(1.1)
   })
 
   describe('Arbitrary', () => {
-    it('generates valid NegativeFloatStrings', () => {
-      validateArbitrary(NegativeFloatString, NegativeFloatString.isNegativeFloatString)
+    it('generates valid PositiveFloatStrings', () => {
+      validateArbitrary(PositiveFloatString, PositiveFloatString.isPositiveFloatString)
     })
   })
 })
